Move conf marker structs to their own module
