+++ eyelid-match-ops/src/primitives/poly/modular_poly/conf.rs
@@ -5,26 +5,10 @@ use std::fmt::Debug;
 use ark_ff::{PrimeField, Zero};
 use lazy_static::lazy_static;
 
-use crate::primitives::poly::Fq79;
+use crate::{primitives::poly::Fq79, FullRes, IrisBits};
 
 #[cfg(tiny_poly)]
-use crate::primitives::poly::fq::FqTiny;
-
-/// The polynomial config used in tests.
-//
-// We use the full resolution by default, but TinyTest when cfg(tiny_poly) is set.
-#[cfg(not(tiny_poly))]
-pub type TestRes = FullRes;
-
-/// The polynomial config used in tests.
-///
-/// Temporarily switch to this tiny field to make test errors easier to debug:
-/// ```no_run
-/// RUSTFLAGS="--cfg tiny_poly" cargo test
-/// RUSTFLAGS="--cfg tiny_poly" cargo bench --features benchmark
-/// ```
-#[cfg(tiny_poly)]
-pub type TestRes = TinyTest;
+use crate::{primitives::poly::fq::FqTiny, TinyTest};
 
 /// Fixed polynomial parameters.
 ///
@@ -51,12 +35,6 @@ pub trait PolyConf: Copy + Clone + Debug + Eq + PartialEq {
     fn coeff_zero() -> &'static Self::Coeff;
 }
 
-/// Iris bit length polynomial parameters.
-///
-/// This uses the full number of iris bits, which gives an upper bound on benchmarks.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct IrisBits;
-
 impl PolyConf for IrisBits {
     const MAX_POLY_DEGREE: usize = crate::IRIS_BIT_LENGTH;
 
@@ -75,14 +53,6 @@ lazy_static! {
 // TODO: try generic_singleton and see if it performs better:
 // <https://docs.rs/generic_singleton/0.5.0/generic_singleton/macro.get_or_init_thread_local.html>
 
-/// Full resolution polynomial parameters.
-///
-/// These are the parameters for full resolution, according to the Inversed Tech report.
-#[cfg(not(tiny_poly))]
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct FullRes;
-
-#[cfg(not(tiny_poly))]
 impl PolyConf for FullRes {
     const MAX_POLY_DEGREE: usize = 2048;
 
@@ -92,13 +62,6 @@ impl PolyConf for FullRes {
         &FQ79_ZERO
     }
 }
-
-/// Tiny test polynomials, used for finding edge cases in tests.
-///
-/// The test parameters are specifically chosen to make failing tests easy to read and diagnose.
-#[cfg(tiny_poly)]
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct TinyTest;
 
 #[cfg(tiny_poly)]
 impl PolyConf for TinyTest {
